🐛 FIX: errors in github actions workflows (trying to write/remove temporary files in tests)

+++ tests/test_data.py
@@ -2,6 +2,7 @@ import os
 import unittest
 
 from easyneuron._testutils import log_errors
+from easyneuron._testutils.gh_actions import notRunningInGitHubActions
 from easyneuron.data import gen_stairs, load_random_humans
 
 
@@ -10,9 +11,11 @@ class TestDataLoading(unittest.TestCase):
 	@log_errors
 	def test_random_humans(self):
 		load_random_humans() # just try to run it to check for errors
-		load_random_humans(filename="logs/temp/" + self.__module__ + ".csv")
 
-		os.remove("logs/temp/" + self.__module__ + ".csv") # remove file from storage
+		if notRunningInGitHubActions():
+			load_random_humans(filename="logs/temp/" + self.__module__ + ".csv")
+
+			os.remove("logs/temp/" + self.__module__ + ".csv") # remove file from storage
 
 class TestDataGen(unittest.TestCase):
